Fix import of a new address (without `--replace` option)
The `execute` method does not have a significant result value. The
`present` variable just contains the cursor object, which is always
`true`. Therefore the else-case for creating a new address entry can't
be reached.

To fix the behavior we need to fetch the result of the select statement.
If the statement doesn't match anything the fetched result will be
`None` which evaluates to `false` and leads to creation of a new address
entry.